[Refactor] 데이터 요청 로직 분리 #1203
