+++ pages/tournament-record.tsx
@@ -1,17 +1,46 @@
-import { useState, useRef, useEffect } from 'react';
-import { SwiperRef } from 'swiper/react';
-import { TournamentInfo } from 'types/tournamentTypes';
+import { useState, useRef, useEffect, useCallback } from 'react';
+import { SwiperClass, SwiperRef } from 'swiper/react';
+import { TournamentInfo, TournamentData } from 'types/tournamentTypes';
+import { instance } from 'utils/axios';
+import { InfiniteScroll } from 'utils/infinityScroll';
 import UserTournamentBracket from 'components/tournament/UserTournamentBracket';
 import LeagueButtonGroup from 'components/tournament-record/LeagueButtonGroup';
 import WinnerSwiper from 'components/tournament-record/WinnerSwiper';
 import WinnerTournamentInfo from 'components/tournament-record/WinnerTournamentInfo';
+import LoadingSpinner from 'components/UI/LoadingSpinner';
 import styles from 'styles/tournament-record/TournamentRecord.module.scss';
 
 export default function TournamentRecord() {
   const [tournamentInfo, setTournamentInfo] = useState<TournamentInfo>();
   const [selectedType, setSelectedType] = useState<string>('ROOKIE');
-  const [isEmpty, setIsEmpty] = useState<boolean>(false);
   const swiperRef = useRef<SwiperRef>(null);
+
+  const fetchTournamentData = useCallback(
+    async (page: number) => {
+      const res = await instance.get(
+        `pingpong/tournaments?page=${page}&type=${selectedType}&size=5&status=END`
+      );
+      return res.data;
+    },
+    [selectedType]
+  );
+
+  const { data, hasNextPage, fetchNextPage, isLoading } =
+    InfiniteScroll<TournamentData>(
+      ['tournamentData', selectedType],
+      fetchTournamentData,
+      'JC01'
+    );
+
+  const indexChangeHandler = useCallback(
+    (swiper: SwiperClass) => {
+      const slidesLength = swiper.slides.length;
+      if (hasNextPage && swiper.activeIndex >= slidesLength - 3) {
+        fetchNextPage();
+      }
+    },
+    [hasNextPage, fetchNextPage]
+  );
 
   useEffect(() => {
     if (swiperRef.current?.swiper) {
@@ -23,19 +52,28 @@ export default function TournamentRecord() {
     <div className={styles.pageWrap}>
       <h1 className={styles.title}>Hall of Fame</h1>
       <LeagueButtonGroup onSelect={setSelectedType} />
-      <WinnerSwiper
-        type={selectedType}
-        size={5}
-        setTournamentInfo={setTournamentInfo}
-        ref={swiperRef}
-        setIsEmpty={setIsEmpty}
-      />
-      {isEmpty ? (
-        <p>종료된 토너먼트가 없습니다!</p>
+      {isLoading ? (
+        <LoadingSpinner />
       ) : (
         <>
-          <WinnerTournamentInfo tournamentInfo={tournamentInfo} />
-          <UserTournamentBracket tournamentId={tournamentInfo?.tournamentId} />
+          <WinnerSwiper
+            tournamentData={data}
+            type={selectedType}
+            size={5}
+            setTournamentInfo={setTournamentInfo}
+            onIndexChange={indexChangeHandler}
+            ref={swiperRef}
+          />
+          {data?.pages[0].totalPage === 0 ? (
+            <p>종료된 토너먼트가 없습니다!</p>
+          ) : (
+            <>
+              <WinnerTournamentInfo tournamentInfo={tournamentInfo} />
+              <UserTournamentBracket
+                tournamentId={tournamentInfo?.tournamentId}
+              />
+            </>
+          )}
         </>
       )}
     </div>
